[0136] 将 read 相关代码从 s7.c 拆分到 s7_scheme_read.c
