+++ devel/0136.md
@@ -1,0 +1,41 @@
+# [0136] 将 read 相关代码从 s7.c 拆分到 s7_scheme_read.c
+
+## 任务相关的代码文件
+- src/s7.c
+- src/s7_scheme_read.c（新增）
+- src/s7_scheme_read.h（新增）
+- src/s7_internal_helpers.h
+- xmake.lua
+
+## 如何测试
+```bash
+xmake b goldfish
+bin/gf tests/scheme/read/read-test.scm
+bin/gf tests/scheme/read-test.scm
+bin/gf bench/read-perf.scm
+```
+
+## 2026-08-20 拆分 read 代码
+
+### What
+1. 新建 `src/s7_scheme_read.c` / `src/s7_scheme_read.h`（与 `s7_scheme_write.c` 对称），从 s7.c 迁移约 740 行 read 相关代码：
+   - port 读原语：`*_read_char`、`*_read_line`、`*_read_semicolon`、`*_read_white_space`、`*_read_name`/`*_read_sharp`（file/string/function/closed_port 等变体）
+   - `read_file()`（0135 优化的整体读入快路径）
+   - Scheme 层函数：`s7_read`、`g_read`、`s7_read_char`、`g_read_char`、`read_char_p_p`、`read_char_chooser`、`s7_peek_char`、`g_peek_char`、`g_read_byte`、`read_line_p_p(p_p)`、`g_read_string`、`input_port_if_not_loading`、`s7i_port_read_line` 等
+2. s7.c 中新增 `s7i_*` 桥接函数（`s7i_token`、`s7i_make_sharp_constant`、`s7i_resize_strbuf`、`s7i_backchar`、`s7i_eval`、`s7i_input_file_functions` 等），声明加入 `s7_internal_helpers.h`；`push_input_port`、`method_or_bust_pp` 去掉 static 导出
+3. port 函数表（`input_file_functions`、`input_string_functions_1` 等）留在 s7.c，通过访问器桥接供 read_file 使用
+4. g_* 函数的 H_/Q_ 文档宏保留在 s7.c（init 的 defun 宏引用它们）
+
+### Why
+s7.c 已有 8.6 万行。继 write 侧拆出 `s7_scheme_write.c` 之后，将 read 侧（端口读原语、reader、read 家族 Scheme 函数）拆出，便于后续 read 性能优化与维护。
+
+### How
+- `s7_scheme_read.c` include `s7_internal.h`（含 `port_t`、`token_t`、opcode 枚举、栈宏、`struct s7_scheme` 完整定义），热路径（`string_read_char`、`string_read_name` 等）直接用本地宏访问 port 字段，无函数调用开销
+- 字符分类表（`char_ok_in_a_name`、`number_table`、`white_space`）来自 `s7_ctables.h` 的 extern 声明
+- 深层依赖（`token()`、`make_sharp_constant`、`eval` 等核心 static）通过 `s7i_*` 包装函数桥接
+- 注意：s7.c 中 `port_data_block(p)` 是 `port_port(p)->block`（port_t 字段），与 `port_block(p)`（cell 的 `object.prt.block`）不同字段，read.c 中的本地宏必须与 s7.c 一致，否则关闭端口时 `free_port_data` 会对空指针 liberate 导致崩溃
+
+### 验证
+- tests/scheme 下 302 个测试全部通过
+- tests/liii、tests/srfi 下 67 个测试全部通过
+- bench/read-perf.scm：读取全部仓库 .scm 文件 10 次迭代 1.73s（0135 基线 2.34s，无回退）
