+++ src/s7_internal_helpers.h
@@ -394,6 +394,22 @@ s7_pointer s7i_multiply_p_ppp_wrapped(s7_scheme *sc, s7_pointer x, s7_pointer y,
 s7_pointer s7i_invert_p_p(s7_scheme *sc, s7_pointer x);
 s7_pointer s7i_divide_p_pp(s7_scheme *sc, s7_pointer x, s7_pointer y);
 
+/* bridge functions for s7_scheme_read.c (reader migration) */
+int32_t s7i_token(s7_scheme *sc);
+s7_pointer s7i_make_sharp_constant(s7_scheme *sc, const char *name, bool with_error, s7_pointer port, bool error_if_bad_number);
+void s7i_resize_strbuf(s7_scheme *sc, s7_int needed_size);
+void s7i_backchar(char c, s7_pointer port);
+bool s7i_is_loader_port(s7_pointer p);
+s7_pointer s7i_an_input_port_string_obj(void);
+s7_pointer s7i_an_open_input_port_string_obj(void);
+s7_pointer s7i_eval(s7_scheme *sc, s7_int op);
+block_t *s7i_mallocate_port(s7_scheme *sc);
+void s7i_port_set_filename(s7_scheme *sc, s7_pointer port, const char *name, size_t len);
+void push_input_port(s7_scheme *sc, s7_pointer new_port);
+s7_pointer method_or_bust(s7_scheme *sc, s7_pointer obj, s7_pointer method, s7_pointer args, s7_pointer typ, int32_t num);
+s7_pointer method_or_bust_p(s7_scheme *sc, s7_pointer obj, s7_pointer method, s7_pointer typ);
+s7_pointer method_or_bust_pp(s7_scheme *sc, s7_pointer obj, s7_pointer method, s7_pointer x1, s7_pointer x2, s7_pointer typ, int32_t num);
+
 #ifdef __cplusplus
 }
 #endif
